Removes unused pops Renames AddressBookEntry with AddressBookEntry
